requirements.txt
asgiref==3.8.1
bcrypt==4.3.0
Django==5.1.7
django-environ==0.12.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
dotenv==0.9.9
psycopg2-binary==2.9.10
PyJWT==2.9.0
python-dotenv==1.0.1
sqlparse==0.5.3
django-cors-headers==3.14.0
drf_yasg==1.21.10
